Added encryption_functions.py
